WIP, improve WMTS/api example
