+++ docs/apilayer.html
@@ -3,4 +3,5 @@ layout: default
 script: gridlayer.js
 title: Gridlayer
 ---
+<p>This WMTS layer autosaves all tiles.</p>
 <div id="map" style="height: 75vh"></div>
