+++ src/TileManager.js
@@ -241,7 +241,7 @@ export async function removeTile(key) {
  * @returns {Promise<Blob>}
  */
 export async function getTile(key) {
-  return (await dbPromise).get(tileStoreName, key).then((result) => result.blob);
+  return (await dbPromise).get(tileStoreName, key).then((result) => result && result.blob);
 }
 
 /**
